splitted remove access component and corrected test names
