+++ opencti-platform/opencti-graphql/tests/03-integration/01-database/playbookComponents/access-restrictions-component-test.ts
